done refactoring
