+++ jbugs-web/src/app/bug/bugs/bugs.component.html
@@ -11,7 +11,6 @@
   class="ag-theme-balham"
   [rowData]="bugList"
   [columnDefs]="columnDefs"
-
   (rowClicked)='onRowClicked($event)'
 >
 </ag-grid-angular>
