--- jbugs-web/src/app/roles/role/role.component.html
@@ -1,3 +1,0 @@
-<p>
-  role works!
-</p>
