--- jbugs-web/src/app/user/user-list/user-list.component.html
@@ -1,12 +1,0 @@
-<div class="test-header">Selection: <span id="selectedRow"></span></div>
-<ag-grid-angular
-  id="myGrid"
-  style="width: 1000px; height: 1000px;"
-  class="ag-theme-balham"
-  [rowData]="userList"
-  [columnDefs]="columnDefs"
-  [rowSelection]="rowSelection"
-  (selectionChanged)="onSelectionChanged()"
-  (gridReady)="onGridReady($event)"
->
-</ag-grid-angular>
