--- jbugs-web/src/app/user/user-list/user-list.component.spec.ts
@@ -1,25 +1,0 @@
-import {async, ComponentFixture, TestBed} from '@angular/core/testing';
-
-import {UserListComponent} from './user-list.component';
-
-describe('UserListComponent', () => {
-  let component: UserListComponent;
-  let fixture: ComponentFixture<UserListComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [UserListComponent]
-    })
-      .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(UserListComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
