--- jbugs-web/src/app/user/user.component.css
@@ -1,3 +1,0 @@
-li {
-  text-align: left;
-}
