+++ jbugs-web/src/app/user/user.module.ts
@@ -1,17 +1,18 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {UserListComponent} from "./user-list/user-list.component";
-import {UserComponent} from "./user.component";
+import {UsersComponent} from "./users/users.component";
 import {JumperDirective} from "./directives/jumper.directive";
 import {RbtnDirective} from "./directives/rbtn.directive";
 import {BackendService} from "./services/backend.service";
 import {AgGridModule} from "ag-grid-angular"
+import {UserEditComponent} from "./user-edit/user-edit.component";
+import {UserAddComponent} from "./user-add/user-add.component";
+import {RouterModule} from "@angular/router";
 
 
 @NgModule({
   declarations: [
-    UserListComponent,
-    UserComponent,
+    UsersComponent,
     JumperDirective,
     RbtnDirective
   ],
@@ -20,11 +21,11 @@ import {AgGridModule} from "ag-grid-angular"
   ],
   imports: [
     CommonModule,
-    AgGridModule.withComponents([])
+    AgGridModule.withComponents([]),
+    RouterModule
   ],
   exports: [
-    UserListComponent,
-    UserComponent,
+    UsersComponent,
     JumperDirective,
     RbtnDirective
   ]
