+++ jbugs-web/src/app/user/users/users.component.css
@@ -1,0 +1,3 @@
+li {
+  text-align: left;
+}
