+++ jbugs-web/src/app/user/users/users.component.ts
@@ -1,27 +1,36 @@
-import {Component, EventEmitter, OnInit, Output} from '@angular/core';
-import {addUser} from '../../models/addUser.model';
-import {UserService} from '../services/user.service';
-import { Router } from '@angular/router';
-import {RestUser} from "../../models/restUser.models";
-
+import {Component, OnInit} from '@angular/core';
+import {UserService} from "../services/user.service";
+import {addUser} from "../../models/addUser.model";
+import {Router} from "@angular/router";
 
 @Component({
-  selector: 'app-users',
+  selector: 'app-user-list',
   templateUrl: './users.component.html',
   styleUrls: ['./users.component.css']
 })
 export class UsersComponent implements OnInit {
 
-  public userList: RestUser[];
-  public currentUser: RestUser;
-  public id: number;
+  private gridApi;
+  private gridColumnApi;
 
+  private rowSelection;
 
-  @Output()
-  public output = new EventEmitter<RestUser>();
-  public showList = true;
+  private columnDefs;
+  public self;
 
-  constructor(private userService: UserService, private router:Router ) {
+  public userList: addUser[];
+
+  constructor(private userService: UserService, private router: Router) {
+    this.columnDefs = [
+      {headerName: 'Firstname', field: 'firstName',sortable:true,filter:true},
+      {headerName: 'Lastname', field: 'lastName',sortable:true,filter:true},
+      {headerName: 'Email', field: 'email',sortable:true,filter:true},
+      {headerName: 'Mobile', field: 'mobileNumber',sortable:true,filter:true},
+      {headerName: 'Username', field: 'username',sortable:true,filter:true}
+
+    ];
+    this.rowSelection="single";
+
   }
 
 
@@ -29,15 +38,28 @@ export class UsersComponent implements OnInit {
     this.userService.getAllUsers()
       .subscribe((userList) => {
         this.userList = userList;
-      });
+      })
   }
 
-  updateUser(person: RestUser) {
-    this.currentUser = person;
-    alert(person.firstName);
-    //this.id = 1;
-    //this.router.navigateByUrl("userid/edit");
-    this.router.navigate([ "/edit" ,this.currentUser.username]);
+  onGridReady(params) {
+    this.gridApi = params.api;
+    this.gridColumnApi = params.columnApi;
   }
 
+  onSelectionChanged() {
+    let selectedRows = this.gridApi.getSelectedRows();
+    let router = this.router;
+    let selectedRowsString = "";
+    selectedRows.forEach(function(selectedRow, index) {
+      if (index !== 0) {
+        selectedRowsString += ", ";
+      }
+      selectedRowsString += selectedRow.firstName;
+      document.getElementById('selectedRow').innerText=selectedRowsString;
+      alert(selectedRow.username);
+      router.navigate(["dashboard/users/edit", selectedRow.username]);
+    });
+
+  }
 }
+
